refactor: userController to singleton. Add style
