+++ src/controllers/appController.ts
@@ -1,9 +1,11 @@
 import RouterController from './routerController';
 import { assertDefined } from '../helpers/helpers';
+import UserController from './userController';
 
 export default class AppController {
     constructor() {
         const router: RouterController = RouterController.getInstance();
+        const userController: UserController = UserController.getInstance();
         const viewContainer: HTMLElement = assertDefined(document.querySelector('.content'));
         router.setRootElement(viewContainer);
         router.reOpenCurrent();
@@ -15,7 +17,7 @@ export default class AppController {
             router.navigate('/registration')
         );
         assertDefined(document.querySelector('#logout')).addEventListener('click', () => router.navigate('/logout'));
-        if (localStorage.getItem('jwt') === null) {
+        if (userController.isSignin()) {
             assertDefined(document.querySelector('#signin')).classList.add('hidden');
             assertDefined(document.querySelector('#registration')).classList.add('hidden');
         } else {
