+++ src/views/signinView.ts
@@ -2,35 +2,25 @@ import UserController from '../controllers/userController';
 import { assertDefined } from '../helpers/helpers';
 import FormService from '../service/formService';
 import ViewInterface from './viewInterface';
-
+import './authView.css';
 class SigninView extends ViewInterface {
     controller: UserController;
     constructor(rootElement: HTMLElement) {
         super(rootElement);
-        this.controller = new UserController();
+        this.controller = UserController.getInstance();
     }
 
     show(): void {
-        const formContainer = document.createElement('div');
-        formContainer.classList.add('form-container');
-        const form = document.createElement('div');
-        form.classList.add('signin-form');
+        const form = FormService.getAuthForm();
         this.fillLoginForm(form);
         form.addEventListener('input', () =>
             assertDefined(form.querySelector<HTMLParagraphElement>('#errMesage')).classList.add('hidden')
         );
-        formContainer.append(form);
         this.rootElement.innerText = '';
-        this.rootElement.append(formContainer);
+        this.rootElement.append(form);
     }
 
     private fillLoginForm(form: HTMLElement) {
-        const errMesage = document.createElement('p');
-        errMesage.innerText = 'Неверный логин или пароль';
-        errMesage.classList.add('errMesage');
-        errMesage.classList.add('hidden');
-        errMesage.id = 'errMesage';
-        form.append(errMesage);
         const email = FormService.getInput(['login-form__login'], 'email', 'email...', 'email');
         form.append(email);
         const pass = FormService.getInput(['login-form__pass'], 'password', 'Пароль...', 'password');
